+++ fe1-web/features/social/network/ReactionHandler.ts
@@ -1,11 +1,10 @@
 import { ExtendedMessage } from 'model/network/method/message';
-import {
-  addReaction, dispatch, getStore, makeCurrentLao,
-} from 'store';
+import { dispatch, getStore, makeCurrentLao } from 'store';
 import { ActionType, MessageRegistry, ObjectType } from 'model/network/method/message/data';
 
 import { Reaction } from '../objects';
 import { AddReaction } from './messages/reaction';
+import { addReaction } from '../reducer/SocialReducer';
 
 /**
  * Handler for social media chirp's reactions
