+++ fe1-web/model/objects/wallet/Token.ts
@@ -1,6 +1,7 @@
 import { derivePath, getPublicKey } from 'ed25519-hd-key';
-import { EventStore, OpenedLaoStore, WalletStore } from 'store';
+import { OpenedLaoStore, WalletStore } from 'store';
 import { RollCall } from 'features/rollCall/objects';
+import { EventStore } from 'features/events/store';
 import { Hash } from '../Hash';
 import { PopToken } from '../PopToken';
 import { Base64UrlData } from '../Base64Url';
